+++ geo-types/CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+## Unreleased
+* Add rstar compatibility for MultiPolygon
+
 ## 0.7.13
 
 * POSSIBLY BREAKING: Minimum supported version of Rust (MSRV) is now 1.70